rename: 도메인 기준으로 디렉토리 재정의 feature/TODAK-197/update-mockup
