+++ app/(main)/my/setting/page.tsx
@@ -1,13 +1,13 @@
 "use client";
 
 import UploadFileLabel from "@/components/my/UploadFileLabel";
-import { RootState } from "@/feature/redux";
+import { RootState } from "@/redux";
 import {
   changeNickname,
   createCharacter,
   fetchMemberInfo,
-} from "@/feature/redux/slices/member/memberExtraReducers";
-import { setNickname } from "@/feature/redux/slices/member/memberSlice";
+} from "@/redux/slices/member/memberExtraReducers";
+import { setNickname } from "@/redux/slices/member/memberSlice";
 import { ChangeNicknameType, CreateCharacterType } from "@/utils/types/dto";
 import { Accordion, Button, HR, Label, TextInput } from "flowbite-react";
 import { useRouter } from "next/navigation";
