+++ domain/auth/slices/index.ts
@@ -1,0 +1,4 @@
+import loginSlice from "./login/loginSlice";
+import signupSlice from "./signup/signupSlice";
+
+export { loginSlice, signupSlice };
